Refactor BtpTransport and related services for improved efficiency and readability
- Removed unused BtpUtils.hpp and integrated utility functions into BinaryUtil.hpp.
- Replaced mutex-based synchronization with MessageQueue for handling asynchronous tasks in BtpTransportScheduler.
- Simplified BtpTransport implementation by removing unnecessary state management and using std::span for data handling.
- Updated BLE service implementations (CurrentScheduleService, UpdateScheduleService, SystemInfoService) to use TLV format for data processing.
- Introduced TlvReader and TlvWriter classes for better handling of TLV encoded data.
- Enhanced error handling and data validation in BLE services.
- Removed legacy code and improved overall code structure for maintainability.

+++ BleServer.hpp
@@ -21,7 +21,7 @@ public:
     template <typename Enum>
         requires std::is_enum_v<Enum>
     void addService(Enum type, BleService* service) const {
-        addService(static_cast<int32_t>(type), service);
+        addService(static_cast<uint8_t>(type), service);
     }
 
     void start() const;
